% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_predict}
\alias{add_predict}
\title{add predicted values or residuals under FE projection}
\usage{
add_predict(model, long, var_name, new_var_name,
  var_name_weights = "omega", force_robust = FALSE, residuals = TRUE)
}
\arguments{
\item{model}{list that describes the model specification}

\item{long}{data.table}

\item{var_name}{name of variable}

\item{new_var_name}{name of new variable with residuals}

\item{var_name_weights}{variable with weights}

\item{force_robust}{logical, force using numerically stable (but possibly slow)
algorithm to fit probit}

\item{residuals}{logical, TRUE = compute residuals, FALSE = compute predicted values}
}
\description{
add predicted values or residuals under FE projection
}
