% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_xtilde}
\alias{add_xtilde}
\title{add xtilde's by reference}
\usage{
add_xtilde(model, long, force_robust = FALSE)
}
\arguments{
\item{model}{list that describes the model specification}

\item{long}{data.table}
}
\description{
add xtilde's by reference
}
