% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_ystar}
\alias{add_ystar}
\title{Add latent index to dataset for given structural parameters.}
\usage{
add_ystar(model, theta, gammaS, gammaR, long = model$links,
  trimming_bound = 0)
}
\arguments{
\item{model}{list that describes the model specification}

\item{theta}{homophily parameter}

\item{gammaS}{vector of sender effects}

\item{gammaR}{vector of receiver effects}

\item{long}{data.table}

\item{trimming_bound}{specify trimming sequence for probabilities}
}
\value{
ystar (=latent index) is added by reference, a view on the data.table is returned
}
\description{
Add latent index to dataset for given structural parameters.
}
