% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{compute_J}
\alias{compute_J}
\title{Compute J (see paper for definition)}
\usage{
compute_J(ystar1, ystar2, rho, r12 = NULL)
}
\arguments{
\item{ystar1}{ystar for ij link}

\item{ystar2}{ystar for ji link}

\item{rho}{correlation (reciprocity parameter)}

\item{r12}{precomputed dyad probability can be supplied for efficiency gains}
}
\value{
vector
}
\description{
Compute J (see paper for definition)
}
