% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{compute_omega}
\alias{compute_omega}
\title{Compute omega for ij link (definition see paper)}
\usage{
compute_omega(ystar, p = compute_p(ystar))
}
\arguments{
\item{ystar}{ystar for ij link}

\item{p}{pre-computed linking probability}
}
\value{
vector
}
\description{
Compute omega for ij link (definition see paper)
}
