% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{compute_p}
\alias{compute_p}
\title{Compute linking probability p for ij link (definition see paper)}
\usage{
compute_p(ystar)
}
\arguments{
\item{ystar}{ystar for ij link}
}
\value{
vector
}
\description{
Compute linking probability p for ij link (definition see paper)
}
