% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{compute_p1}
\alias{compute_p1}
\title{Compute p(1 - p) for ij link (definition see paper)}
\usage{
compute_p1(ystar, p = compute_p(ystar))
}
\arguments{
\item{ystar}{ystar for ij link}

\item{p}{pre-computed linking probability}
}
\value{
vector
}
\description{
Compute p(1 - p) for ij link (definition see paper)
}
