% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{compute_r}
\alias{compute_r}
\title{Compute probability of reciprocated link}
\usage{
compute_r(ystar1, ystar2, rho)
}
\arguments{
\item{ystar1}{ystar for ij link}

\item{ystar2}{ystar for ji link}

\item{rho}{correlation (reciprocity parameter)}
}
\value{
vector of probabilities
}
\description{
Compute probability of reciprocated link
}
