% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_ystar}
\alias{compute_ystar}
\title{Compute ystar}
\usage{
compute_ystar(links, theta, gammaS, gammaR, X_names, col_sender,
  col_receiver, num_nodes, trimming_bound = 0)
}
\arguments{
\item{links}{data.table with link data}

\item{theta}{theta}

\item{gammaS}{gammaS}

\item{gammaR}{gammaR}

\item{X_names}{column(s) of regressors}

\item{col_sender}{column that gives sender}

\item{col_receiver}{column that gives receiver}

\item{num_nodes}{number of nodes}

\item{trimming_bound}{specify trimming sequence for probabilities}
}
\value{
ystar vector
}
\description{
Compute ystar
}
