% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_dyadic_data.R
\name{copy_var_to_new_suffix}
\alias{copy_var_to_new_suffix}
\title{Copy variable, new name changes the suffix}
\usage{
copy_var_to_new_suffix(dt, variables, old_suffix, new_suffix)
}
\arguments{
\item{dt}{data.table}

\item{variables}{name of variables (columns)}

\item{old_suffix}{old suffix}

\item{new_suffix}{new suffix}
}
\value{
view of data.table (new variables are added by reference)
}
\description{
Copy variable, new name changes the suffix
}
