% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model.R
\name{define_model}
\alias{define_model}
\title{Define a network model}
\usage{
define_model(d, X_names = NULL, col_sender = "i", col_receiver = "j",
  col_outcome = "Y")
}
\arguments{
\item{d}{data.frame with link data}

\item{X_names}{covariate names (NULL = no covariates)}

\item{col_sender}{name of column that identifies sender}

\item{col_receiver}{name of column that identifies receiver}
}
\value{
list describing the network model and data
}
\description{
Define a network model
}
