% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{define_sim_design_jochmans}
\alias{define_sim_design_jochmans}
\title{Define a simulation design (Jochmans)}
\usage{
define_sim_design_jochmans(N, theta, rho, type_C_N = "0")
}
\arguments{
\item{N}{number of nodes}

\item{theta}{coefficient for parametric part}

\item{rho}{within dyad correlation}

\item{type_C_N}{sparsity parameter
* "0"
* "loglog"
* "sqrt_log"
* "log"}
}
\value{
simulation design as list
}
\description{
Define a simulation design (Jochmans)
}
