% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_triangles.R
\name{double_bootstrap}
\alias{double_bootstrap}
\title{Double bootstrap procedure}
\usage{
double_bootstrap(wide, rho_bs0, model, bs_iterations = 200,
  alpha = 0.1, trimming_bound = 0, bias_estimator = "theorem2",
  force_robust = FALSE)
}
\arguments{
\item{wide}{link data in wide format}

\item{rho_bs0}{within-dyad correlation of bootstrap distribution}

\item{model}{model description in special list format}

\item{bs_iterations}{number of bootstrap replications}

\item{alpha}{significance level}

\item{trimming_bound}{bound for trimming small probabilities}

\item{bias_estimator}{estimator of bias (see triangle test)}

\item{force_robust}{logical, force using numerically stable (but possibly slow)
algorithm to fit probit}
}
\value{
list with bootstrapped SE and quantiles
}
\description{
Double bootstrap procedure
}
