% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_triangles.R
\name{double_bootstrap_inner}
\alias{double_bootstrap_inner}
\title{Fast computation of bootstrap bias}
\usage{
double_bootstrap_inner(wide, rho_bs0, model, bs_iterations = 200,
  trimming_bound = 0)
}
\arguments{
\item{wide}{link data in wide format}

\item{rho_bs0}{within-dyad correlation of bootstrap distribution}

\item{model}{model description in special list format}

\item{bs_iterations}{number of bootstrap replications}

\item{trimming_bound}{bound for trimming small probabilities}
}
\value{
bootstrap sample (vector)
}
\description{
Fast computation of bootstrap bias
}
