% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_triangles.R
\name{draw_bootstrap_sample}
\alias{draw_bootstrap_sample}
\title{Draw from bootstrap distribution}
\usage{
draw_bootstrap_sample(wide, rho_bs0, model, values_ystar_ij,
  values_ystar_ji, col_y_bs = "Y_bs")
}
\arguments{
\item{wide}{link data wide format (has to contain i and j, canonically ordered)}

\item{rho_bs0}{within-dyad correlation of bootstrap distribution}

\item{model}{model description in special list format}

\item{values_ystar_ij}{of bootstrap process}

\item{values_ystar_ji}{of bootstrap process}

\item{col_y_bs}{column name for bootstrapped Y}
}
\value{
data.table bootstrap sample
}
\description{
Draw from bootstrap distribution
}
