% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{draw_j18_dynamic_meeting}
\alias{draw_j18_dynamic_meeting}
\title{Sample from alternative with dynacmic behavior}
\usage{
draw_j18_dynamic_meeting(design)
}
\arguments{
\item{design}{network parameters}
}
\value{
list containing the generated data and fixed effects
}
\description{
Sample from alternative with dynacmic behavior
}
