% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{draw_network_jochmans_2018}
\alias{draw_network_jochmans_2018}
\title{Sample a random network following modified Jochmans design}
\usage{
draw_network_jochmans_2018(design)
}
\arguments{
\item{design}{network parameters}
}
\value{
list containing the generated data and fixed effects
}
\description{
Sample a random network following modified Jochmans design
}
