% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{excess_trans}
\alias{excess_trans}
\title{compute observed excess transitivity}
\usage{
excess_trans(df, y_variable, p_variable)
}
\arguments{
\item{df}{data frame with link data (ordered canonically)}

\item{y_variable}{string that contains name of column that stores link indicator}

\item{p_variable}{string that contains name of column that stores link probability}
}
\value{
number of expected transitive triangles
}
\description{
assumes that link observations are ordered canonically a
}
