% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_triangles.R
\name{fast_excess_transitivity}
\alias{fast_excess_transitivity}
\title{Compute excess transitivity}
\usage{
fast_excess_transitivity(links, col_link_indicator, col_link_probability,
  col_sender = "i", col_receiver = "j")
}
\arguments{
\item{links}{data frame with link data}

\item{col_link_indicator}{name of column with binary link indicator}

\item{col_link_probability}{name of column with link probability}

\item{col_sender}{name of column that gives sender id}

\item{col_receiver}{name of column that gives receiver id}
}
\value{
list with number of transitive triangles, expected triangles, excess transitivity
}
\description{
count transitive triangles and compute excess transitivity at the same time
for efficiency gain
}
