% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_dyadic_data.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Reshape dyadic network data from long (ij, i != j) to wide (ij, i<j)}
\usage{
long_to_wide(links, col_sender, col_receiver, upper_triangle = TRUE,
  nodes = unique(links[[col_sender]]),
  ij_variables = setdiff(colnames(links), c(col_sender, col_receiver)))
}
\arguments{
\item{links}{network data in long format}

\item{col_sender}{name of column containing the sender of the link}

\item{col_receiver}{name of column containing the receiver of the link}

\item{upper_triangle}{=TRUE (default) wide format respects i<j,
=FALSE then i != j but each row contains information about everything in dyad}

\item{nodes}{all node indices}

\item{ij_variables}{variables that will be reported as dyad information}
}
\value{
network data in wide format
}
\description{
Reshape dyadic network data from long (ij, i != j) to wide (ij, i<j)
}
