% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_estimation.R
\name{MLE_stage1}
\alias{MLE_stage1}
\title{ML estimation of stage 1}
\usage{
MLE_stage1(model, start = NULL, robust = FALSE)
}
\arguments{
\item{model}{list describing the model (generated by define_model)}

\item{start}{optional start values for theta (currently ignored)}

\item{robust}{flag, if TRUE don't use speedglm}
}
\value{
fitted model
}
\description{
ML estimation of stage 1
}
