% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_estimation.R
\name{MLE_stage2}
\alias{MLE_stage2}
\title{ML estimation stage 2}
\usage{
MLE_stage2(fit_MLE_stage1, gammaS = fit_MLE_stage1$gammaS_hat,
  gammaR = fit_MLE_stage1$gammaR_hat, method = "mle", rho_start = 0,
  kappa = 0.02)
}
\arguments{
\item{fit_MLE_stage1}{returned list from stage 1}

\item{gammaS}{sender effects}

\item{gammaR}{receiver effects}

\item{method}{maximize log likelihood ("mle") or search for root of gradient ("gradient")}

\item{rho_start}{starting value of rho (reciprocity parameter)}

\item{kappa}{bound rho away from -1 and 1 (see paper)}
}
\value{
fitted model
}
\description{
ML estimation stage 2
}
