% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_estimation.R
\name{MLE_stage2_gradient}
\alias{MLE_stage2_gradient}
\title{Stage 2: solve likelihood directly}
\usage{
MLE_stage2_gradient(wide, bound_lower, bound_upper, rho_start = 0,
  inital_search_radius = 0.4,
  convergence_tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{wide}{network data in upper-triangular wide format}

\item{bound_lower}{lower bound on rho}

\item{bound_upper}{upper bound on rho}

\item{rho_start}{starting value (default = 0)}

\item{inital_search_radius}{intial search radius around starting value (default = 0.4)}
}
\value{
fitted model
}
\description{
Stage 2: solve likelihood directly
}
