% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_estimation.R
\name{MLE_stage2_mle}
\alias{MLE_stage2_mle}
\title{Stage 2: solve likelihood directly}
\usage{
MLE_stage2_mle(wide, bound_lower, bound_upper, rho_start = 0)
}
\arguments{
\item{wide}{network data in upper-triangular wide format}

\item{bound_lower}{lower bound on rho}

\item{bound_upper}{upper bound on rho}

\item{rho_start}{starting value (default = 0)}
}
\value{
fitted model
}
\description{
Stage 2: solve likelihood directly
}
