% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_pairs}
\alias{num_pairs}
\title{compute number of (ordered) pairs from 1, ..., N}
\usage{
num_pairs(N)
}
\arguments{
\item{N}{number of elements}
}
\value{
number of ordered pairs
}
\description{
compute number of (ordered) pairs from 1, ..., N
}
