% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_trans_closures}
\alias{num_trans_closures}
\title{Detect number of transitive relationships closed by this link}
\usage{
num_trans_closures(df, y_variable)
}
\arguments{
\item{df}{data frame with link data (ordered canonically)}

\item{y_variable}{string that contains name of column that stores link indicator}
}
\value{
number of supporting transitive triangles for this link
}
\description{
assumes that link observations are ordered canonically a
}
