% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{partial_p_ystar}
\alias{partial_p_ystar}
\title{Compute derivative of p (linking probability) wrt to ystar}
\usage{
partial_p_ystar(ystar)
}
\arguments{
\item{ystar}{index}
}
\value{
vector of same length as ystar
}
\description{
Compute derivative of p (linking probability) wrt to ystar
}
