% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{partial_r_ystar1}
\alias{partial_r_ystar1}
\title{Compute derivative of r wrt rho}
\usage{
partial_r_ystar1(ystar1, ystar2, rho)
}
\arguments{
\item{ystar1}{ystar for ij link}

\item{ystar2}{ystar for ji link}

\item{rho}{correlation (reciprocity parameter)}
}
\value{
vector
}
\description{
Compute derivative of r wrt rho
}
