% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{partial_r_ystar1_ystar2}
\alias{partial_r_ystar1_ystar2}
\title{Compute 2nd order derivative of r wrt ystar1 and ystar2}
\usage{
partial_r_ystar1_ystar2(ystar1, ystar2, rho)
}
\arguments{
\item{ystar1}{ystar for ij link}

\item{ystar2}{ystar for ji link}

\item{rho}{correlation (reciprocity parameter)}
}
\value{
vector
}
\description{
Compute 2nd order derivative of r wrt ystar1 and ystar2
}
