% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_link_data}
\alias{read_link_data}
\title{use C++ class LinkVariable to establish link to R data frame}
\usage{
read_link_data(df, i, j, var_name)
}
\arguments{
\item{df}{data frame with link data (ordered canonically)}

\item{i}{Sender index}

\item{j}{Receiver index}

\item{var_name}{name of variabe in data frame}
}
\value{
var_name at i, j
}
\description{
this function is mainly for testing to see if the link is 
established correctly
}
