% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_methods}
\alias{simulate_methods}
\title{Run simulations}
\usage{
simulate_methods(sim_design, draw_network, num_simulations = 100,
  bs_iterations = 100, stage2_method = "gradient",
  triangles_bias_est = "theorem2", triangles_bs_inf = "none",
  file_name = NULL, verbose = TRUE, bs_trimming_bound = 0,
  force_robust = FALSE, bs_force_robust = force_robust)
}
\arguments{
\item{sim_design}{Simulation design in list format}

\item{draw_network}{function that draws a network (takes argument sim_design)}

\item{num_simulations}{number of simulations}

\item{bs_iterations}{number of bootstrap iterations for triangle test}

\item{stage2_method}{either "mle" or "gradient"}

\item{triangles_bias_est}{specify how bias is estimated}

\item{triangles_bs_inf}{specify typoe of bootstrap inference}

\item{file_name}{file name for saving results (relative to path)
if not provided then results are not saved to disk}

\item{verbose}{print more diagnostic output}

\item{bs_trimming_bound}{specify trimming of probabilities in bootstrap Monte Carlo}

\item{force_robust}{don't use speedglm (for robustness in sparse networks)}

\item{bs_force_robust}{don't use speedglm in bootstrap Monte Carlo}
}
\value{
data frame with simulation results
}
\description{
Run simulations
}
