% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_triangles.R
\name{triangle_test}
\alias{triangle_test}
\title{Compute specification test based on transitive triangles}
\usage{
triangle_test(fit_ML, fit_ttest_theta, gammaS = fit_ML$gammaS_hat,
  gammaR = fit_ML$gammaR_hat, theta = fit_ML$theta_hat,
  rho = fit_ML$rho_hat_MLE, rho0 = 0, bs_iterations = 200,
  alpha = 0.1, bias_estimator = "theorem2", bs_inference = "none",
  trimming_bound = 0, bs_trimming_bound = trimming_bound,
  force_robust = FALSE, bs_force_robust = force_robust,
  compute_other_variances = FALSE)
}
\arguments{
\item{fit_ML}{returned object from ML estimation (\code{MLE_stage1})}

\item{fit_ttest_theta}{fitted \code{ttest_theta} object}

\item{gammaS}{vector of sender effects  (default: estimated value in \code{fit_ML})}

\item{gammaR}{vetor of receiver effects (default: estimated value in \code{fit_ML})}

\item{theta}{value used for \eqn{\theta} (default: estimated value in \code{fit_ML})}

\item{rho}{value used for \eqn{\rho} (default: estimated value in \code{fit_ML});
not needed if both \code{bias_estimator = "theorem2"} and \code{bs_inference = "none"}}

\item{rho0}{value of theta under null hypothesis
(default: zero/test significance)}

\item{bs_iterations}{number of bootstrap iterations}

\item{alpha}{alpha used to compute bootstrap critical values}

\item{bias_estimator}{specify estimator of bias term
\itemize{
  \item \code{"theorem1"} based on theorem in paper, needs estimate of rho
  \item \code{"theorem2"} based on theorem in paper, doesn't need estimate of rho
  \item \code{"boostrap"} boostrap the bias
  }}

\item{bs_inference}{specify what kind of bootstrap is used to determine whether the test rejects
\itemize{
  \item \code{"none"} no bootstrap inference
  \item \code{"double"} double bootstrap
  }}

\item{trimming_bound}{specify trimming of probabilities used to compute statistics}

\item{bs_trimming_bound}{specify trimming of probabilities used in
simulation of bootstrap distribution}

\item{force_robust}{logical, don't use speedglm}

\item{bs_force_robust}{logical, don't use speedglm in simulation of bootstrap distribution}

\item{compute_other_variances}{logical, force computation of standard errors for infeasible
triangle tests}
}
\value{
list with test statistic, rejection and auxilary output
}
\description{
Compute specification test based on transitive triangles
}
