% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triangles_biascorr_new_vars}
\alias{triangles_biascorr_new_vars}
\title{add new link variables that are required for bias correction}
\usage{
triangles_biascorr_new_vars(df)
}
\arguments{
\item{df}{data frame with link data (ordered canonically)}
}
\value{
data frame with new variables
}
\description{
assumes that link observations are ordered canonically and
that data frame contains the variables "p", "H", "partial_p"
}
