% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_rho.R
\name{ttest_rho}
\alias{ttest_rho}
\title{Compute bias adjusted t-statistic for parameter rho}
\usage{
ttest_rho(fit_ML, fit_ttest_theta, gammaS = fit_ML$gammaS_hat,
  gammaR = fit_ML$gammaR_hat, theta = fit_ML$theta_hat,
  rho = fit_ML$rho_hat, rho0 = 0, var_meat_estimator = "theorem",
  trimming_bound = 0, force_robust = FALSE)
}
\arguments{
\item{fit_ML}{returned object from ML estimation}

\item{fit_ttest_theta}{fitted ttest_theta object}

\item{gammaS}{value used for gammaS (default: estimated value)}

\item{gammaR}{value used for gammaR (default: estimated value)}

\item{theta}{value used for theta (default: estimated value)}

\item{rho}{value used for rho (default: estimated value)}

\item{rho0}{value of theta under null hypothesis
(default: zero/test significance)}

\item{var_meat_estimator}{method to compute "meat" of the variance sandwich
* "theorem" implement exactly as in theorem (put hats on everything)
* "clustering" clustering based on first-order expansion (will not depend on rho_hat! )}

\item{trimming_bound}{specify trimming of probabilities}

\item{force_robust}{don't use speedglm (for robustness in sparse networks)}
}
\value{
list with t statistic and auxilary output
}
\description{
Compute bias adjusted t-statistic for parameter rho
}
