% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_theta.R
\name{ttest_theta}
\alias{ttest_theta}
\title{Compute bias adjusted t-statistic for parameter theta}
\usage{
ttest_theta(fit_ML, gammaS = fit_ML$gammaS_hat,
  gammaR = fit_ML$gammaR_hat, theta = fit_ML$theta_hat,
  rho = fit_ML$rho_hat, theta0 = rep.int(0, length(theta)),
  var_meat_estimator = "clustering", trimming_bound = 0,
  force_robust = FALSE)
}
\arguments{
\item{fit_ML}{returned object from ML estimation}

\item{gammaS}{value used for gammaS (default: estimated value)}

\item{gammaR}{value used for gammaR (default: estimated value)}

\item{theta}{value used for theta (default: estimated value)}

\item{rho}{value used for rho (default: estimated value)}

\item{theta0}{value of theta under null hypothesis
(default: zero vector/test significance)}

\item{var_meat_estimator}{method to compute "meat" of the variance sandwich
* "theorem" implement exactly as in theorem (put hats on everything)
* "clustering" clustering based on first-order expansion (will not depend on rho_hat! )}

\item{trimming_bound}{specify trimming of probabilities}

\item{force_robust}{don't attempt to use speedglm}
}
\value{
list with t statistic and auxilary output
}
\description{
Compute bias adjusted t-statistic for parameter theta
}
