% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_inner_prod}
\alias{weighted_inner_prod}
\title{Compute a weighted inner product}
\usage{
weighted_inner_prod(A, B, weights)
}
\arguments{
\item{A}{matrix (n x k)}

\item{B}{matrix (n x k)}

\item{weights}{vector of weights (1 x n)}
}
\value{
inner product weighted by w
}
\description{
Compute a weighted inner product
}
