% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_probabilities.R
\name{partial_J_ystar1}
\alias{partial_J_ystar1}
\title{Compute derivative of J (see paper) wrt ystar1}
\usage{
partial_J_ystar1(ystar1, ystar2, rho, r = NULL, r_rho = NULL,
  r_ystar1 = NULL, r_ystar1_rho = NULL)
}
\arguments{
\item{ystar1}{ystar for ij link}

\item{ystar2}{ystar for ji link}

\item{rho}{correlation (reciprocity parameter)}
}
\value{
vector
}
\description{
Compute derivative of J (see paper) wrt ystar1
}
